split substrate account selector in the separate component
