+++ packages/widget/src/components/index.ts
@@ -2,5 +2,6 @@ export { AmountSelector } from './amount-selector';
 export { NetworkSelector } from './network-selector';
 export { OverlayComponent } from './common/overlay-component';
 export { ConnectWalletButton } from './common/buttons/connect-wallet';
+export { SubstrateAccountSelector } from './substrate-account-selector';
 export { AddressInput } from './address-input';
 export { FungibleTokenTransfer } from './transfer/fungible/fungible-token-transfer';
