+++ packages/widget/src/components/substrate-account-selector/index.ts
@@ -1,0 +1,1 @@
+export { SubstrateAccountSelector } from './substrate-account-selector';
